SPG/requirements.txt
django
django-rest-framework
psycopg2
psycopg2-binary
djangorestframework-simplejwt=5.2.2
python-dotenv=1.0.0

